update: http untils
